+++ examples/ts/myScene.ts
@@ -2,9 +2,9 @@
 import DuckEngine, { Duck } from '../../dist';
 
 export default class MainScene extends DuckEngine.Scene {
-	public myRect: Duck.GameObjects.GameObject;
-	public myCircle: Duck.GameObjects.GameObject;
-	public gameObjects: Duck.Group<Duck.GameObjects.GameObject>;
+	public myRect: Duck.TypeClasses.GameObjects.GameObject;
+	public myCircle: Duck.TypeClasses.GameObjects.GameObject;
+	public gameObjects: Duck.Group<Duck.TypeClasses.GameObjects.GameObject>;
 
 	public myCamera: Duck.Camera;
 
@@ -21,10 +21,11 @@ export default class MainScene extends DuckEngine.Scene {
 		this.myRect = this.add.gameobject.rect(0, 0, 15, 15, '#fff');
 		this.myCircle = this.add.gameobject.circle(50, 50, 10, '#fff');
 
-		this.gameObjects = this.add.group<Duck.GameObjects.GameObject>(
-			'gameObjects',
-			[this.myRect, this.myCircle]
-		);
+		this.gameObjects =
+			this.add.group<Duck.TypeClasses.GameObjects.GameObject>(
+				'gameObjects',
+				[this.myRect, this.myCircle]
+			);
 
 		// camera
 		this.myCamera = this.add.mainCamera();