+++ src/core/effect/preset/explosion.ts
@@ -1,8 +1,8 @@
 import { Duck } from '../../../index';
 import Game from '../../game';
 import randomFloat from '../../math/randomFloat';
-import Particle from '../../particles/particle';
-import ParticleEmitter from '../../particles/particleEmitter';
+import Particle from '../../gameobjects/particles/particle';
+import ParticleEmitter from '../../gameobjects/particles/particleEmitter';
 import Scene from '../../scene';
 import Effect from '../effect';