+++ src/core/math/randomFloat.ts
@@ -1,5 +1,3 @@
-import { Duck } from '../../index';
-
 /**
  * @function
  * @description Returns a random float between two ranges